Task 6

+++ 0x00-hello_world/6-size.c
@@ -2,17 +2,17 @@
 /**
  * main - Entry
  *
- * Description: output size
+ * Description: output sizes of data types
  *
  * Return: Always 0 (success)
  *
 */
 int main(void)
 {
-	printf("size of a char: %i byte(s)\n", sizeof(char));
-	printf("size of an int: %i byte(s)\n", sizeof(int));
-	printf("size of a long int: %i byte(s)\n", sizeof(long int));
-	printf("size of a long long int: %i byte(s)\n", sizeof(long long int));
-	printf("size of a float int: %i byte(s)\n", sizeof(float));
+	printf("size of a char: %lu byte(s)\n",(unsigned long)sizeof(char));
+	printf("size of an int: %lu byte(s)\n",(unsigned long)sizeof(int));
+	printf("size of a long int: %lu byte(s)\n",(unsigned long)sizeof(long int));
+	printf("size of a long long int: %lu byte(s)\n",(unsigned long)sizeof(long long int));
+	printf("size of a float int: %lu byte(s)\n",(unsigned long)sizeof(float));
 	return (0);
 }
